Switched to using the system-property file.encoding where files are read/written
SCL-3087 :
Passing the name of a codepage to the constructor of ParseUnit now sets the system-property file.encoding. This is not really used because the encoding must be set when starting the JVM. However this property can be set and the new value can be read. This is now used where files have previously been read using FileReaders to instead use InputStreamReaders reading FileInputStreams with the Charset defined by the system-property file.encoding.