Throwing Exceptions for incorrect codepage names
The constructor ParseUnit(File file, String codepage) now throws different exceptions for different errors in the codepage name. IllegalCharsetNameException if it contains illegal characters or was empty, UnsupportedCharsetException if it is not recognized by the system and IllegalArgumentException if the codepage was null. There are also new unit-test to ensure that those exceptions are thrown as expected.